api-gateway/requirements.txt
Flask-RESTful
PyJWT
Flask-SQLAlchemy
requests
psycopg2-binary